doc, tools: extract toHTML
